refactor(storage): split tencent_cos driver into submodules
Split the monolithic tencent_cos.rs driver (755 lines) into a multi-file module structure for better code organization:

- `mod.rs`: Core driver struct, validation, and trait implementations (110 lines)
- `signing.rs`: COS signature generation, URL construction, and encoding utilities (203 lines)
- `native_preview.rs`: Document preview feature using COS CI (76 lines)
- `native_thumbnail.rs`: Image thumbnail generation with COS CI (110 lines)
- `tests.rs`: All unit tests moved to dedicated test module (299 lines)

No functional changes. All existing features, APIs, and test coverage preserved.
